refactor(workspace-manager): organize menu buttons under menu-bar subfolder
