+++ src/features/workspace-manager/components/menu-bar/WorkspaceRenameButton/WorkspaceRenameButton.tsx
@@ -1,4 +1,4 @@
-import MenuBarItem from "../../../../components/MenuBarItem/MenuBarItem";
+import MenuBarItem from "../../../../../components/MenuBarItem/MenuBarItem";
 
 export default function WorkspaceRenameButton() {
   const handleClick = () => {};
